Move the platform info stuff to utils/platform.
