+++ src/client/common/platform/types.ts
@@ -4,35 +4,10 @@
 import * as fs from 'fs';
 import * as semver from 'semver';
 import { Disposable } from 'vscode';
+import { Architecture, OSDistro, OSType } from '../utils/platform';
 
-export enum Architecture {
-    Unknown = 1,
-    x86 = 2,
-    x64 = 3
-}
-export enum OSType {
-    Unknown,
-    Windows,
-    OSX,
-    Linux
-}
-export enum OSDistro {
-    Unknown,
-    // linux:
-    Ubuntu,
-    Debian,
-    RHEL,
-    Fedora,
-    CentOS,
-    // The remainder aren't officially supported.
-    // See: https://code.visualstudio.com/docs/supporting/requirements
-    Suse,
-    Gentoo,
-    Arch
-}
-
-export const IOSInfo = Symbol('IOSInfo');
-export interface IOSInfo {
+export const IPlatformInfo = Symbol('IPlatformInfo');
+export interface IPlatformInfo {
     readonly type: OSType;
     readonly arch: string;
     readonly version: semver.SemVer;
@@ -53,16 +28,15 @@ export interface IRegistry {
 
 export const IPlatformService = Symbol('IPlatformService');
 export interface IPlatformService {
-    os: IOSInfo;
-    pathVariableName: 'Path' | 'PATH';
-    virtualEnvBinName: 'bin' | 'scripts';
+    readonly info: IPlatformInfo;
+    readonly pathVariableName: 'Path' | 'PATH';
+    readonly virtualEnvBinName: 'bin' | 'scripts';
 
-    // tslint:disable-next-line: no-suspicious-comment
-    // TODO: Drop the following (in favor of osType).
-    isWindows: boolean;
-    isMac: boolean;
-    isLinux: boolean;
-    is64bit: boolean;
+    // convenience methods
+    readonly isWindows: boolean;
+    readonly isMac: boolean;
+    readonly isLinux: boolean;
+    readonly is64bit: boolean;
 }
 
 export type TemporaryFile = { filePath: string } & Disposable;
