+++ src/client/common/utils/version.ts
@@ -5,6 +5,17 @@
 
 import * as semver from 'semver';
 
+export function parseVersion(raw: string): semver.SemVer {
+    raw = raw.replace(/\.00*(?=[1-9]|0\.)/, '.');
+    const ver = semver.coerce(raw);
+    if (ver === null || !semver.valid(ver)) {
+        // tslint:disable-next-line: no-suspicious-comment
+        // TODO: Raise an exception instead?
+        return new semver.SemVer('0.0.0');
+    }
+    return ver;
+}
+
 export function convertToSemver(version: string) {
     const versionParts = (version || '').split('.').filter(item => item.length > 0);
     while (versionParts.length < 3) {
