+++ src/client/interpreter/locators/services/windowsRegistryService.ts
@@ -3,8 +3,9 @@ import { inject, injectable } from 'inversify';
 import * as _ from 'lodash';
 import * as path from 'path';
 import { Uri } from 'vscode';
-import { Architecture, IRegistry, RegistryHive } from '../../../common/platform/types';
+import { IRegistry, RegistryHive } from '../../../common/platform/types';
 import { IPathUtils, Is64Bit } from '../../../common/types';
+import { Architecture } from '../../../common/utils/platform';
 import { IServiceContainer } from '../../../ioc/types';
 import { IInterpreterHelper, InterpreterType, PythonInterpreter } from '../../contracts';
 import { CacheableLocatorService } from './cacheableLocatorService';
