+++ src/test/common/variables/envVarsService.test.ts
@@ -4,9 +4,9 @@
 import { expect, use } from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 import * as path from 'path';
-import { isWindows } from '../../../client/common/platform/osinfo';
 import { PathUtils } from '../../../client/common/platform/pathUtils';
 import { IPathUtils } from '../../../client/common/types';
+import { isWindows } from '../../../client/common/utils/platform';
 import { EnvironmentVariablesService } from '../../../client/common/variables/environment';
 import { IEnvironmentVariablesService } from '../../../client/common/variables/types';
