+++ src/test/install/channelManager.channels.test.ts
@@ -8,8 +8,8 @@ import { QuickPickOptions } from 'vscode';
 import { IApplicationShell } from '../../client/common/application/types';
 import { InstallationChannelManager } from '../../client/common/installer/channelManager';
 import { IModuleInstaller } from '../../client/common/installer/types';
-import { Architecture } from '../../client/common/platform/types';
 import { Product } from '../../client/common/types';
+import { Architecture } from '../../client/common/utils/platform';
 import { IInterpreterLocatorService, InterpreterType, PIPENV_SERVICE, PythonInterpreter } from '../../client/interpreter/contracts';
 import { ServiceContainer } from '../../client/ioc/container';
 import { ServiceManager } from '../../client/ioc/serviceManager';
