+++ src/test/install/channelManager.messages.test.ts
@@ -7,8 +7,9 @@ import * as TypeMoq from 'typemoq';
 import { IApplicationShell } from '../../client/common/application/types';
 import { InstallationChannelManager } from '../../client/common/installer/channelManager';
 import { IModuleInstaller } from '../../client/common/installer/types';
-import { Architecture, IPlatformService } from '../../client/common/platform/types';
+import { IPlatformService } from '../../client/common/platform/types';
 import { Product } from '../../client/common/types';
+import { Architecture } from '../../client/common/utils/platform';
 import { IInterpreterService, InterpreterType, PythonInterpreter } from '../../client/interpreter/contracts';
 import { ServiceContainer } from '../../client/ioc/container';
 import { ServiceManager } from '../../client/ioc/serviceManager';
