+++ src/test/install/pythonInstallation.unit.test.ts
@@ -7,8 +7,9 @@ import { Container } from 'inversify';
 import * as TypeMoq from 'typemoq';
 import { IApplicationShell } from '../../client/common/application/types';
 import { PythonInstaller } from '../../client/common/installer/pythonInstallation';
-import { Architecture, IFileSystem, IPlatformService } from '../../client/common/platform/types';
+import { IFileSystem, IPlatformService } from '../../client/common/platform/types';
 import { IPersistentStateFactory, IPythonSettings } from '../../client/common/types';
+import { Architecture } from '../../client/common/utils/platform';
 import { IInterpreterHelper, IInterpreterLocatorService, IInterpreterService, InterpreterType, PythonInterpreter } from '../../client/interpreter/contracts';
 import { InterpreterHelper } from '../../client/interpreter/helpers';
 import { ServiceContainer } from '../../client/ioc/container';
