+++ src/test/unittests/pytest/pytest.testparser.unit.test.ts
@@ -7,9 +7,8 @@ import { expect, use } from 'chai';
 import * as chaipromise from 'chai-as-promised';
 import * as typeMoq from 'typemoq';
 import { CancellationToken, OutputChannel, Uri } from 'vscode';
-import { getOSType } from '../../..//client/common/platform/osinfo';
 import { IApplicationShell, ICommandManager } from '../../../client/common/application/types';
-import { OSType } from '../../../client/common/platform/types';
+import { getOSType, OSType } from '../../../client/common/utils/platform';
 import { IServiceContainer } from '../../../client/ioc/types';
 import { TestsHelper } from '../../../client/unittests/common/testUtils';
 import { TestFlatteningVisitor } from '../../../client/unittests/common/testVisitors/flatteningVisitor';
